+++ frontend/lib/main.dart
@@ -1,145 +1,48 @@
 // Autor: Vinícius Castro & Allan Giovanni Matias Paes
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:frontend/firebase_options.dart';
+import 'package:frontend/pages/forgot_password_page.dart';
+import 'package:frontend/pages/home_page.dart';
+import 'package:frontend/pages/login_page.dart';
+import 'package:frontend/pages/register_page.dart';
 
-void main() {
-  runApp(const MyApp());
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  try {
+    await dotenv.load(fileName: ".env");
+  } catch (e) {
+    // .env might not exist or be empty, we can continue if not strictly needed for all environments
+    print("Warning: Could not load .env file: $e");
+  }
+
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+  runApp(const MesclaInvest());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class MesclaInvest extends StatelessWidget {
+  const MesclaInvest({super.key});
 
   @override
   Widget build(BuildContext context) {
+    var userName = FirebaseAuth.instance.currentUser?.displayName;
+
     return MaterialApp(
+      title: 'Mescla Invest',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.green),
+        colorScheme: ColorScheme.fromSeed(seedColor: const Color(0xFF00A84E)),
         useMaterial3: true,
-        appBarTheme: const AppBarTheme(
-          backgroundColor: Colors.white,
-          foregroundColor: Colors.black,
-          elevation: 0,
-        ),
+        fontFamily: 'Inter', // Assuming Inter or standard sans-serif
       ),
-      home: const ForgotPasswordPage(),
-    );
-  }
-}
-
-class ForgotPasswordPage extends StatefulWidget {
-  const ForgotPasswordPage({super.key});
-
-  @override
-  State<ForgotPasswordPage> createState() => _ForgotPasswordPageState();
-}
-
-class _ForgotPasswordPageState extends State<ForgotPasswordPage> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _emailController = TextEditingController();
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        leading: const BackButton(color: Colors.black),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(24.0),
-        child: Form(
-          key: _formKey,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Align(
-                alignment: Alignment.centerLeft,
-                child: Container(
-                  width: 60,
-                  height: 45,
-                  decoration: BoxDecoration(
-                    color: const Color(0xFFE8F5E9),
-                    borderRadius: BorderRadius.circular(16),
-                  ),
-                  child: const Icon(
-                    Icons.email_outlined,
-                    color: Color(0xFF2E7D32),
-                    size: 32,
-                  ),
-                ),
-              ),
-              const SizedBox(height: 24),
-              const Text(
-                'Esqueceu a senha?',
-                style: TextStyle(
-                  fontSize: 26,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.black,
-                ),
-              ),
-              const SizedBox(height: 8),
-              const Text(
-                'Digite seu email e enviaremos um código para redefinir sua senha',
-                style: TextStyle(
-                  fontSize: 14,
-                  color: Colors.grey,
-                ),
-              ),
-              const SizedBox(height: 32),
-              const Text(
-                'Email',
-                style: TextStyle(
-                  fontSize: 14,
-                    color: Color.fromARGB(255, 63, 62, 62),
-                ),
-              ),
-              const SizedBox(height: 8),
-              Align(
-                alignment: Alignment.centerLeft,
-                child: SizedBox(
-                  width: 300, // Defina aqui a largura que preferir
-                  child: TextFormField(
-                    controller: _emailController,
-                    decoration: const InputDecoration(
-                      hintText: 'seu@email.com',
-                      border: OutlineInputBorder(),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 24),
-              Align(
-                alignment: Alignment.centerLeft,
-                child: SizedBox(
-                  width: 300,
-                  child: ElevatedButton(
-                    onPressed: () {},
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: const Color(0xFF2E7D32),
-                      foregroundColor: Colors.white,
-                      padding: const EdgeInsets.symmetric(vertical: 16),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                    ),
-                    child: const Text(
-                      'Enviar código',
-                      style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
+      home: HomePage(userName: userName ?? 'Desconhecido'),
+      routes: {
+        '/login': (context) => const LoginPage(),
+        '/register': (context) => const RegisterPage(),
+        '/forgotPassword': (context) => const ForgotPasswordPage()
+      },
     );
   }
 }
